refactor: Eliminate OAuth token store duplication (195 lines removed)
Create BaseOAuthTokenStore abstract class to eliminate duplicated token
management logic across File and Memory OAuth token stores.

Changes:
- Create base-oauth-token-store.ts with shared implementations
- Extract getToken(), findByRefreshToken(), deleteToken(), cleanup()
- MemoryOAuthTokenStore: 173 → 79 lines (54% reduction)
- FileOAuthTokenStore: 360 → 259 lines (28% reduction)
- Update duplication baseline: 350 → 347 clones (-0.9%)

Impact:
- 195 lines eliminated across 2 token stores
- Resolves SonarCloud 85-88% duplication in OAuth token stores
- Maintains 100% test coverage (all validation checks pass)

Technical Details:
- Base class provides hook methods for storage-specific behavior
- onTokenMutated() allows file store to trigger saves
- isExpired() allows custom expiry logic per store
- Full backward compatibility maintained
